requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==2.3.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.7
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==2.1.2
channels-redis==2.3.2
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==2.2.5
Django==2.2.24
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==0.5.6
packaging==21.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
zope.interface==5.4.0
